+++ docs/snippets/trace/component.py
@@ -2,7 +2,7 @@ import asyncio
 
 from grelmicro import Grelmicro
 from grelmicro.log import Log
-from grelmicro.trace import Trace, TracingExporterType, instrument
+from grelmicro.trace import Trace, TraceExporterType, instrument
 
 
 @instrument
@@ -15,7 +15,7 @@ micro = Grelmicro(
         Log(),
         Trace(
             service_name="orders",
-            exporter=TracingExporterType.CONSOLE,
+            exporter=TraceExporterType.CONSOLE,
         ),
     ]
 )